Update display mode colorings
