+++ plugins/alignments/src/shared/components/ColorByTag.tsx
